flask_app/pyproject.toml
[tool.poetry]
name = "flask_app"
version = "0.1.0"
description = ""
authors = ["Muskelo <titanioman@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.1.0"
Flask-SQLAlchemy = "^2.5.1"
mysql-connector-python = "^8.0.28"
Flask-Migrate = "^3.1.0"
Flask-RESTful = "^0.3.9"
pydantic = "^1.9.0"
passlib = "^1.7.4"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
